+++ packages/bpk-component-card/stories.js
@@ -1,17 +1,23 @@
 import React from 'react';
-import { storiesOf, action } from '@kadira/storybook';
+import { storiesOf } from '@kadira/storybook';
 
 import BpkCard from './index';
 
 storiesOf('bpk-component-card', module)
   .add('Default', () => (
-    <BpkCard href="#" onClick={action('normal clicked')}>
+    <BpkCard>
+      Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum
+      sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
+    </BpkCard>
+  ))
+  .add('With href', () => (
+    <BpkCard href="#">
       Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum
       sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
     </BpkCard>
   ))
   .add('Without padding', () => (
-    <BpkCard href="#" onClick={action('no padding clicked')} padded={false}>
+    <BpkCard padded={false}>
       Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum
       sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.
     </BpkCard>